misc fixes: * fix players under the affect of pain spike, wouldn’t take extra damage from evil omen * fix paralyze spell wouldnt work against targets with higher than 110.1 magic resist * Curse spell OSI accurate:  duration should be refreshed when a new curse spell is cast and it would expire further in the future, else it should fizzle when the targets current curse affliction is more powerful * fix mana vampire wouldn’t drain the maximum amount of mana from the target, if the caster had a large amount of mana remaining

+++ Projects/UOContent/Spells/Fifth/Paralyze.cs
@@ -39,7 +39,7 @@ namespace Server.Spells.Fifth
 
                 if (Core.AOS)
                 {
-                    var secs = (int)(GetDamageSkill(Caster) / 10 - GetResistSkill(m) / 10);
+                    var secs = (GetDamageSkill(Caster) / 10 - GetResistSkill(m) / 10);
 
                     if (!Core.AOS)
                     {
